+++ react/admin/OrganizationsList/OrganizationsListSearch/OrganizationsListSearch.tsx
@@ -1,8 +1,10 @@
 import React from 'react'
 import { Search, csx, useSearchState } from '@vtex/admin-ui'
 import type { TagProps } from '@vtex/admin-ui'
+import { useIntl } from 'react-intl'
 
 import type { FetchListOptions } from '../../../organizations/hooks'
+import { organizationMessages as adminMessages } from '../../utils/messages'
 
 export const TagVariantByStatus: Record<string, TagProps['variant']> = {
   active: 'green',
@@ -15,7 +17,9 @@ interface Props {
 }
 
 const OrganizationsListSearch: React.FC<Props> = ({ onSearch }) => {
+  const { formatMessage } = useIntl()
   const search = useSearchState()
+  const { onClear, ...inputProps } = search.getInputProps()
 
   const handleSearchKeyDown: React.KeyboardEventHandler<HTMLFormElement> = event => {
     if (event.key !== 'Enter') return
@@ -25,8 +29,6 @@ const OrganizationsListSearch: React.FC<Props> = ({ onSearch }) => {
       page: 1,
     })
   }
-
-  const { onClear, ...inputProps } = search.getInputProps()
 
   const handleSearchClear = () => {
     onClear()
@@ -41,6 +43,8 @@ const OrganizationsListSearch: React.FC<Props> = ({ onSearch }) => {
     <Search
       rel=""
       {...inputProps}
+      title={formatMessage(adminMessages.searchPlaceholder)}
+      placeholder={formatMessage(adminMessages.searchPlaceholder)}
       onClear={handleSearchClear}
       onKeyDown={handleSearchKeyDown}
       className={csx({ marginRight: '$space-5' })}